% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car_data}
\alias{car_data}
\title{Car Data}
\format{A data frame with 100 rows and 3 variables
\describe{
\item{Color}{The color of the car. These are characters}
\item{Speed}{The maximum speed of the car in m/s}
\item{State}{The state where the car was made}
}}
\usage{
car_data
}
\description{
This dataset contains information of the color, speed and state of cars.
}
\keyword{datasets}
